[REFACTOR] 코드 수정
