+++ src/main/java/store/ckin/coupon/config/DataSourceConfig.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import store.ckin.coupon.config.properties.DbProperties;
 import store.ckin.coupon.keymanager.KeyManager;
 
 import javax.sql.DataSource;
